keeping score!!
but it's splitting attackers/defenders between everyone in the room right now
